Add Fragment for chat

+++ app/src/main/java/com/aibles/zulipeventtest/EventRegisterResponse.kt
@@ -9,5 +9,5 @@ data class EventRegisterResponse(
     @SerializedName("zulip_version") var zulipVersion: String,
     @SerializedName("zulip_feature_level") var zulipFeatureLevel: Int,
     @SerializedName("max_message_id") var maxMessageId: Long,
-    @SerializedName("last_event_id") var lastEventId: Int
+    @SerializedName("last_event_id") var lastEventId: Long
 )
